+++ src/main/java/org/prgms/voucher/VoucherController.java
@@ -1,12 +1,10 @@
 package org.prgms.voucher;
 
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
 
-import java.util.List;
-import java.util.UUID;
-
-@RequestMapping("/api/v1")
-@RestController
+@Controller
 public class VoucherController {
     private final VoucherService voucherService;
 
@@ -15,37 +13,13 @@ public class VoucherController {
     }
 
     @GetMapping("/vouchers")
-    public List<Voucher> findVouchers() {
-        return voucherService.getVouchers();
+    public String vouchersPage(Model model) {
+        var vouchers = voucherService.getVouchers();
+        System.out.println(vouchers);
+        model.addAttribute("vouchers", vouchers);
+        return "voucher-list";
     }
 
-    @GetMapping("/vouchers/{voucherId}")
-    public Voucher findVoucher(@PathVariable("voucherId") UUID voucherId) {
-        return voucherService.getVoucher(voucherId);
-    }
-
-    @DeleteMapping("/voucher/{voucherId}")
-    public void deleteVoucher(@PathVariable("voucherId") UUID voucherId) {
-        System.out.println("ssss");
-        voucherService.deleteVoucher(voucherId);
-    }
-
-    @GetMapping("/vouchers/type/{voucherType}")
-    public List<Voucher> findVouchersByType(@PathVariable("voucherType") VoucherType voucherType) {
-        return voucherService.getVouchersByType(voucherType);
-    }
-
-    @PostMapping("/voucher/FIXED_AMOUNT")
-    public Voucher addFixedVoucher(@RequestBody CreateVoucherRequest createVoucherRequest) {
-        var voucher = voucherService.createVoucher(VoucherType.FIXED_AMOUNT, createVoucherRequest.getAmount());
-        return voucher;
-    }
-
-    @PostMapping("/voucher/PERCENT_DISCOUNT")
-    public Voucher addPercentVoucher(@RequestBody CreateVoucherRequest createVoucherRequest) {
-        var voucher = voucherService.createVoucher(VoucherType.PERCENT_DISCOUNT, createVoucherRequest.getAmount());
-        return voucher;
-    }
 
 
 }
